remove redirect for AddUser
