+++ src/pages/addOrder.js
@@ -1,10 +1,12 @@
 import React from "react";
 import { Redirect } from "react-router-dom";
 
-import { editOrderControl } from "../components/protectedRoute/permissions";
-import { isUser } from "../components/protectedRoute/permissions";
-import LoginSubComponent from "../components/loginSubComponent";
 import AddView from "../components/addView";
+import {
+  editOrderControl,
+  isUser,
+} from "../components/protectedRoute/permissions";
+import LoginSubComponent from "../components/loginSubComponent";
 
 const AddOrder = () => {
   return isUser() ? (
